requirements.txt
discord.py
yt-dlp
requests
python-dotenv
PyNaCl
